Set undefined variables globally
The behaviour for defined variables (either preassigned or loop
variables) is to update their value; this doesn't change.

The old behaviour for undefined variables was to set their value in the
scope of the innermost loop. An example where this makes a difference is
for the code `for(1)Assign('a',k);`, which previously only set `k` in
the scope of the `for` loop, so the value would be lost after the loop.

The rework of input variables means that they are no longer preassigned,
so this change restores the old behaviour where they were set globally.